rag_agent/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
pydantic-settings==2.6.1
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
langchain==1.2.6
langchain-core==1.2.7
langchain-openai==1.1.7
openai==2.15.0
pgvector==0.2.5
python-dotenv==1.0.1
